fixed resource module
