+++ projects/flyout/package.json
@@ -8,7 +8,7 @@
     "@angular/forms": "^12.2.15",
     "@angular/router": "^12.2.15",
     "@skyux/core": "^5.2.2",
-    "@skyux/i18n": "^5.0.1",
+    "@skyux/i18n": "^5.0.0-beta.0",
     "@skyux/indicators": "^5.1.2",
     "@skyux/router": "^5.0.2",
     "@skyux/theme": "^5.0.4"
