--- projects/flyout/src/modules/shared/flyout-resources.module.ts
@@ -1,39 +1,0 @@
-/**
- * NOTICE: DO NOT MODIFY THIS FILE!
- * The contents of this file were automatically generated by
- * the 'ng generate @skyux/i18n:lib-resources-module' schematic.
- * To update this file, simply rerun the command.
- */
-
- import { NgModule } from '@angular/core';
- import {
-   getLibStringForLocale,
-   SkyAppLocaleInfo,
-   SkyI18nModule,
-   SkyLibResources,
-   SkyLibResourcesProvider,
-   SKY_LIB_RESOURCES_PROVIDERS
- } from '@skyux/i18n';
-
- const RESOURCES: { [locale: string]: SkyLibResources } = {
-   'EN-US': {"skyux_flyout_close":{"message":"Close flyout"},"skyux_flyout_iterator_previous_button":{"message":"Previous"},"skyux_flyout_iterator_next_button":{"message":"Next"},"skyux_flyout_permalink_button":{"message":"View record"},"skyux_flyout_primary_action_button":{"message":"Create list"},"skyux_flyout_resize_handle":{"message":"Resize the flyout window. Press the enter or space key to activate, then use left arrow key to expand the flyout window and right arrow key to shrink the flyout window. Press the enter, space, or tab key to deactivate."}},
- };
-
- export class FlyoutResourcesProvider implements SkyLibResourcesProvider {
-   public getString(localeInfo: SkyAppLocaleInfo, name: string): string {
-     return getLibStringForLocale(RESOURCES, localeInfo.locale, name);
-   }
- }
-
- /**
-  * Import into any component library module that needs to use resource strings.
-  */
- @NgModule({
-   exports: [SkyI18nModule],
-   providers: [{
-     provide: SKY_LIB_RESOURCES_PROVIDERS,
-     useClass: FlyoutResourcesProvider,
-     multi: true
-   }]
- })
- export class SkyFlyoutResourcesModule { }
